Add my own caching system as Serenity's one is broken for members
Discord does not seem to send the guild members on bot startup, or
Serenity fails to recieve those members and to cache them.

To be able to run some of the bot commands in a timely manner, the cache
is a definite necessity, thus the addition of a custom cache system for
users and members.