bot: fix reaction handler leading to internal error spam
When a reaction is added in a channel where the bot has permission to
receive reactions but _not_ read messages, it would error out trying to
fetch the message contents. Discord permissions are weird.

This is fixed by finding a started game first, then fetching the
message, base on the assumption that the bot will have permission to
read messages in a channel where a game is running.